+++ src/environments/definitions/optimism/environment.ts
@@ -13,6 +13,7 @@ const createEnvironment = (
   rpcUrls?: string[],
   indexerUrl?: string,
   governanceIndexerUrl?: string,
+  lunarIndexerUrl?: string,
 ) =>
   createEnvironmentConfig({
     key: "optimism",
@@ -20,13 +21,15 @@ const createEnvironment = (
     chain: {
       ...optimism,
       rpcUrls: {
-        default: { http: rpcUrls || optimism.rpcUrls.default.http },
+        default: { http: rpcUrls || ["https://rpc.moonwell.fi/main/evm/10"] },
       },
     },
     transport: rpcUrls
       ? fallback(rpcUrls.map((url) => http(url)))
-      : http(optimism.rpcUrls.default.http[0]),
+      : http("https://rpc.moonwell.fi/main/evm/10"),
     indexerUrl: indexerUrl || "https://ponder.moonwell.fi",
+    lunarIndexerUrl:
+      lunarIndexerUrl || "https://lunar-services-worker.moonwell.workers.dev",
     governanceIndexerUrl:
       governanceIndexerUrl ||
       "https://lunar-services-worker.moonwell.workers.dev",